requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.15.0
pydantic>=2.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
pytest>=7.0.0
requests-mock>=1.9.0